% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_activities.R
\name{dc_activities}
\alias{dc_activities}
\title{DataCite REST API: activities}
\usage{
dc_activities(
  ids = NULL,
  query = NULL,
  limit = 25,
  page = 1,
  cursor = NULL,
  ...
)
}
\arguments{
\item{ids}{(character) one or more activity IDs}

\item{query}{(character) Query string}

\item{limit}{(numeric/integer) results per page}

\item{page}{(numeric/integer) the page to get results for. default: 1}

\item{cursor}{(character) page cursor (used instead of \code{limit} param).
to use cursor pagination, set \code{cursor = 1}, then use the link in
\verb{$links$next}}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
DataCite REST API: activities
}
\details{
for more info on the \verb{/activities} route see
https://support.datacite.org/docs/tracking-provenance
}
\examples{
\dontrun{
if (dc_check()) {
x <- dc_activities()
x
# dc_activities(x$data$id[1]) # FIXME: doesn't work, returns no data
# dc_activities(query = "ecology") # FIXME: this thlimit a 500 error
}}
}
