% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midden.R
\docType{data}
\name{dc_caching}
\alias{dc_caching}
\alias{dc_midden}
\title{Caching control}
\format{An object of class \code{midden} (inherits from \code{R6}) of length 11.}
\usage{
dc_midden
}
\description{
Caching control
}
\examples{
\dontrun{
dc_midden
dc_midden$
}
}
\keyword{datasets}
