% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_client_prefixes.R
\name{dc_client_prefixes}
\alias{dc_client_prefixes}
\title{DataCite REST API: client prefixes}
\usage{
dc_client_prefixes(
  query = NULL,
  year = NULL,
  client_id = NULL,
  prefix_id = NULL,
  sort = NULL,
  include = NULL,
  limit = 25,
  page = 1,
  cursor = NULL,
  ...
)
}
\arguments{
\item{query}{(character) Query string}

\item{year}{(integer/numeric/character) a year}

\item{client_id}{a client ID}

\item{prefix_id}{a prefix ID}

\item{sort}{(character) variable to sort by}

\item{include}{(character) vector of fields to return}

\item{limit}{(numeric/integer) results per page}

\item{page}{(numeric/integer) the page to get results for. default: 1}

\item{cursor}{(character) page cursor (used instead of \code{limit} param).
to use cursor pagination, set \code{cursor = 1}, then use the link in
\verb{$links$next}}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
DataCite REST API: client prefixes
}
\examples{
\dontrun{
if (dc_check()) {
x <- dc_client_prefixes()
x
}}
}
