% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_data_center.R
\name{dc_data_center}
\alias{dc_data_center}
\title{DataCite REST API: data center}
\usage{
dc_data_center(data_center_id, include = NULL, ...)
}
\arguments{
\item{data_center_id}{(character) a data center id}

\item{include}{(character) vector of member fields to return}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DataCite REST API: data center
}
\examples{
\dontrun{
dc_data_center("cdl.dryad")
dc_data_center("cdl.dryad",
  include = c("member", "registration-agency"))
}
}
\seealso{
\code{\link[=dc_data_centers]{dc_data_centers()}}

Other rest api methods: \code{\link{dc_data_centers}},
  \code{\link{dc_members}}, \code{\link{dc_member}},
  \code{\link{dc_works}}, \code{\link{dc_work}}
}
