% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_events.R
\name{dc_events}
\alias{dc_events}
\title{DataCite REST API: events}
\usage{
dc_events(
  ids = NULL,
  query = NULL,
  subj_id = NULL,
  obj_id = NULL,
  doi = NULL,
  orcid = NULL,
  prefix = NULL,
  subtype = NULL,
  subject = NULL,
  source_id = NULL,
  registrant_id = NULL,
  relation_type_id = NULL,
  issn = NULL,
  publication_year = NULL,
  year_month = NULL,
  include = NULL,
  sort = NULL,
  limit = 25,
  page = 1,
  cursor = NULL,
  ...
)
}
\arguments{
\item{ids}{(character) one or more event IDs}

\item{query}{(character) Query string}

\item{subj_id}{(character)}

\item{obj_id}{(character)}

\item{doi}{(character)}

\item{orcid}{(character)}

\item{prefix}{(character)}

\item{subtype}{(character)}

\item{subject}{(character)}

\item{source_id}{(character)}

\item{registrant_id}{(character)}

\item{relation_type_id}{(character)}

\item{issn}{(character)}

\item{publication_year}{(character)}

\item{year_month}{(character)}

\item{include}{(character) vector of fields to return}

\item{sort}{(character) variable to sort by}

\item{limit}{(numeric/integer) results per page}

\item{page}{(numeric/integer) the page to get results for. default: 1}

\item{cursor}{(character) page cursor (used instead of \code{limit} param).
to use cursor pagination, set \code{cursor = 1}, then use the link in
\verb{$links$next}}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
DataCite REST API: events
}
\examples{
\dontrun{
if (dc_check()) {
x <- dc_events()
x
dc_events(query = "birds")
# dc_events(x$data$id[1]) # FIXME: doesn't work
# dc_events(x$data$id[1:3]) # FIXME: doesn't work
}}
}
