\name{dc_identify}
\alias{dc_identify}
\title{identify}
\usage{
  dc_identify(transform = TRUE,
    url = "http://oai.datacite.org/oai")
}
\arguments{
  \item{transform}{transform metadata to list (TRUE/FALSE)}

  \item{url}{the base url for DataCite (leave to default)}
}
\description{
  Learn about the Dryad OAI-PMH service
}
\examples{
\dontrun{
dc_identify()
}
}

