% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_member.R
\name{dc_member}
\alias{dc_member}
\title{DataCite REST API: member}
\usage{
dc_member(member_id, ...)
}
\arguments{
\item{member_id}{(character) Member that the data center is associated with}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DataCite REST API: member
}
\examples{
\dontrun{
dc_member("cern")
}
}
\seealso{
\code{\link[=dc_members]{dc_members()}}

Other rest api methods: \code{\link{dc_data_centers}},
  \code{\link{dc_data_center}}, \code{\link{dc_members}},
  \code{\link{dc_works}}, \code{\link{dc_work}}
}
\concept{rest api methods}
