% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_members.R
\name{dc_members}
\alias{dc_members}
\title{DataCite REST API: members}
\usage{
dc_members(query = NULL, member_type = NULL, region = NULL,
  year = NULL, offset = 0, rows = 25, ...)
}
\arguments{
\item{query}{(character) Query string, e.g., data center name}

\item{member_type}{(character) member type, one of "allocating"
or "non-allocating"}

\item{region}{(character) Geographic region, one of "amer", "apac",
or "emea"}

\item{year}{(numeric/integer) year member joined DataCite}

\item{offset}{(numeric/integer) result offset, the record to start at}

\item{rows}{(numeric/integer) results per page}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DataCite REST API: members
}
\examples{
\dontrun{
dc_members(query = "ands")
}
}
\seealso{
\code{\link[=dc_member]{dc_member()}}

Other rest api methods: \code{\link{dc_data_centers}},
  \code{\link{dc_data_center}}, \code{\link{dc_member}},
  \code{\link{dc_works}}, \code{\link{dc_work}}
}
\concept{rest api methods}
