% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dc_oai_getrecord.R
\name{dc_oai_getrecord}
\alias{dc_oai_getrecord}
\title{Retrieve an individual record from the DataCite repository}
\usage{
dc_oai_getrecord(id = NULL, transform = TRUE)
}
\arguments{
\item{id}{DataCite identifier, e.g., "56225"}

\item{transform}{transform metadata to list (TRUE/FALSE)}
}
\description{
Retrieve an individual record from the DataCite repository
}
\examples{
\dontrun{
out <- dc_oai_getrecord("56225", T)
out$metadata # get $identifier, $datestamp, $setSpec, or $metadata
oaih_transform(out$metadata) # transform only metadata to a list

temp <- dc_listidentifiers(from = '2012-07-27') # get some identifiers
temp[[1,"identifier"]] # gets the identifier
dc_oai_getrecord(temp[[1,"identifier"]]) # get the record
}
}

