% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dc_oai_identify}
\alias{dc_oai_identify}
\title{Learn about the Dryad OAI-PMH service}
\usage{
dc_oai_identify(transform = TRUE)
}
\arguments{
\item{transform}{transform metadata to list (TRUE/FALSE)}
}
\description{
Learn about the Dryad OAI-PMH service
}
\examples{
\dontrun{
dc_oai_identify()
}
}

