% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dc_oai_listsets}
\alias{dc_oai_listsets}
\title{Retrieve the set structure of DataCite, useful for selective harvesting}
\usage{
dc_oai_listsets()
}
\description{
Retrieve the set structure of DataCite, useful for selective harvesting
}
\examples{
\dontrun{
datacite_sets <- dc_oai_listsets()
head(datacite_sets)
}
}

