% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_prefixes.R
\name{dc_prefixes}
\alias{dc_prefixes}
\title{DataCite REST API: prefixes}
\usage{
dc_prefixes(include = NULL, limit = 25, page = 1, cursor = NULL, ...)
}
\arguments{
\item{include}{(character) vector of fields to return}

\item{limit}{(numeric/integer) results per page}

\item{page}{(numeric/integer) result page, the record to start at}

\item{cursor}{(character) page cursor (used instead of \code{limit} param)}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DataCite REST API: prefixes
}
\examples{
\dontrun{
if (dc_check()) {
x <- dc_prefixes()
x
dc_prefixes(limit = 3)
}}
}
