% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_reports.R
\name{dc_reports}
\alias{dc_reports}
\title{DataCite REST API: reports}
\usage{
dc_reports(
  ids = NULL,
  platform = NULL,
  report_name = NULL,
  report_id = NULL,
  release = NULL,
  created = NULL,
  created_by = NULL,
  include = NULL,
  limit = 25,
  page = 1,
  ...
)
}
\arguments{
\item{ids}{(character) one or more report IDs}

\item{platform}{(character) x}

\item{report_name}{(character) x}

\item{report_id}{(character) x}

\item{release}{(character) x}

\item{created}{(character) x}

\item{created_by}{(character) x}

\item{include}{(character) vector of fields to return}

\item{limit}{(numeric/integer) results per page}

\item{page}{(numeric/integer) result page, the record to start at}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DataCite REST API: reports
}
\examples{
\dontrun{
x <- dc_reports()
x
dc_reports(limit = 3)
# dc_reports(ids = x$reports$id[1:3]) # FIXME: doesn't work
}
}
