% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dc_search}
\alias{dc_facet}
\alias{dc_mlt}
\alias{dc_search}
\alias{dc_stats}
\title{Use the DataCite RESTful API for search, facet, stats, and mlt queries.}
\usage{
dc_search(..., callopts = list())

dc_facet(..., callopts = list())

dc_stats(..., callopts = list())

dc_mlt(..., callopts = list())
}
\arguments{
\item{...}{Solr parameters passed on to the respective \code{solr} package
function.}

\item{callopts}{Further args passed on to \code{\link[httr]{GET}}}
}
\description{
Use the DataCite RESTful API for search, facet, stats, and mlt queries.
}
\details{
See the \code{solr} package documentation for available parameters. For each
\code{dc_*} function see the equivalently named function in \code{solr}:

\itemize{
 \item dc_search - solr_search
 \item dc_facet - solr_facet
 \item dc_stats - solr_stats
 \item dc_mlt - solr_mlt
}
}
\examples{
\donttest{
# Search
## search for wind, retrieve only doi and title, and return (at max.) 5 results
dc_search(q = "wind", fl=c('doi','title'), rows=5)

## search with csv output
dc_search(q = 'wind', fl=c('doi','title'), wt='csv')

# Faceting
dc_facet(q = "wind", facet.field='publisher')

# Stats
dc_stats(q = "ecology", stats.field='date')

# More like this (aka mlt)
dc_mlt(q = "ecology", mlt.fl='title', mlt.count=5, fl=c('doi','title'))
}
}

