% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_status.R
\name{dc_status}
\alias{dc_status}
\title{DataCite REST API: status of the API}
\usage{
dc_status(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DataCite REST API: status of the API
}
\examples{
\dontrun{
if (dc_check()) {
dc_status()
}}
}
