% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_dataciteCNBaseUrl}
\alias{get_dataciteCNBaseUrl}
\title{Get content negotiation base URL}
\usage{
get_dataciteCNBaseUrl()
}
\value{
contant negotiation base URL
}
\description{
Get content negotiation base URL
}
\seealso{
Other package_customizable_settings: 
\code{\link{package_settings}}
}
\concept{package_customizable_settings}
