% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdatacite-package.R
\docType{package}
\name{rdatacite-package}
\alias{rdatacite-package}
\alias{rdatacite}
\title{DataCite R client.}
\description{
DataCite R client.
}
\section{REST API functions}{

\itemize{
\item \code{\link[=dc_activities]{dc_activities()}} - xxx
\item \code{\link[=dc_client_prefixes]{dc_client_prefixes()}} - xxx
\item \code{\link[=dc_clients]{dc_clients()}} - xxx
\item \code{\link[=dc_dois]{dc_dois()}} - xxx
\item \code{\link[=dc_events]{dc_events()}} - xxx
\item \code{\link[=dc_prefixes]{dc_prefixes()}} - xxx
\item \code{\link[=dc_provider_prefixes]{dc_provider_prefixes()}} - xxx
\item \code{\link[=dc_providers]{dc_providers()}} - xxx
\item \code{\link[=dc_reports]{dc_reports()}} - xxx
\item \code{\link[=dc_status]{dc_status()}} - xxx
}
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
