% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{set_dataciteCNBaseUrl}
\alias{set_dataciteCNBaseUrl}
\title{Set content negotiation base url}
\usage{
set_dataciteCNBaseUrl(url = "https://data.crosscite.org", force = FALSE)
}
\arguments{
\item{url}{A \code{character}. Set the base URL to content negotiation base url.}

\item{force}{A \code{boolean}. Default FALSE.}
}
\description{
Set content negotiation base url
}
