% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dc_oai_getrecord.R
\name{dc_oai_getrecord}
\alias{dc_oai_getrecord}
\title{Retrieve an individual record from the DataCite repository}
\usage{
dc_oai_getrecord(id = NULL)
}
\arguments{
\item{id}{DataCite identifier, e.g., "56225"}
}
\description{
Retrieve an individual record from the DataCite repository
}
\examples{
\dontrun{
dc_oai_getrecord("56225")

temp <- dc_oai_listidentifiers(from = '2012-07-27') # get some identifiers
temp[[1,"identifier"]] # gets the identifier
dc_oai_getrecord(temp[[1,"identifier"]]) # get the record
}
}

