% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_oai_identify.R
\name{dc_oai_identify}
\alias{dc_oai_identify}
\title{Learn about the Dryad OAI-PMH service}
\usage{
dc_oai_identify(...)
}
\arguments{
\item{...}{Curl options passed on to \code{httr::GET}}
}
\description{
Learn about the Dryad OAI-PMH service
}
\examples{
library(crul)
url <- "https://oai.datacite.org/oai"
if (crul::HttpClient$new(url=url)$get()$success()) {
  dc_oai_identify()
}
}
