% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_oai_identify.R
\name{dc_oai_identify}
\alias{dc_oai_identify}
\title{Learn about the Dryad OAI-PMH service}
\usage{
dc_oai_identify(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Learn about the Dryad OAI-PMH service
}
\examples{
\dontrun{
dc_oai_identify()
}
}

