% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dc_oai_listmetadataformats.R
\name{dc_oai_listmetadataformats}
\alias{dc_oai_listmetadataformats}
\title{List available metadata formats that the DataCite repository can disseminate}
\usage{
dc_oai_listmetadataformats(id = NULL, transform = TRUE)
}
\arguments{
\item{id}{DataCite identifier, e.g., "56225"}

\item{transform}{transform metadata to list (TRUE/FALSE)}
}
\description{
List available metadata formats that the DataCite repository can disseminate
}
\examples{
\dontrun{
dc_oai_listmetadataformats()
dc_oai_listmetadataformats("56225")
}
}

