% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_works.R
\name{dc_works}
\alias{dc_works}
\title{DataCite REST API: works}
\usage{
dc_works(query = NULL, order = NULL, sort = NULL, rows = NULL,
  offset = NULL, include = NULL, ...)
}
\arguments{
\item{query}{(character) Query string, e.g., any part of a work}

\item{order}{(character) set the sort order to one of "asc" or "desc"}

\item{sort}{(character) sort results by a certain field, one of
"deposited", "published", "updated" or "score"}

\item{rows}{(numeric/integer) results per page}

\item{offset}{(numeric/integer) result offset, the record to start at}

\item{include}{(character) vector of member fields to return}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DataCite REST API: works
}
\examples{
\dontrun{
dc_works(query = "climate change")
}
}
\seealso{
\code{\link[=dc_work]{dc_work()}}

Other rest api methods: \code{\link{dc_data_centers}},
  \code{\link{dc_data_center}}, \code{\link{dc_members}},
  \code{\link{dc_member}}, \code{\link{dc_work}}
}
\concept{rest api methods}
