% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdatacite-package.R
\docType{package}
\name{rdatacite-package}
\alias{rdatacite-package}
\alias{rdatacite}
\title{DataCite R client.}
\description{
DataCite R client.
}
\section{OAI-PMH functions}{

\itemize{
\item \code{\link[=dc_oai_getrecord]{dc_oai_getrecord()}} - Get records
\item \code{\link[=dc_oai_identify]{dc_oai_identify()}} - identify the OAI-PMH service
\item \code{\link[=dc_oai_listidentifiers]{dc_oai_listidentifiers()}} - List identifiers
\item \code{\link[=dc_oai_listmetadataformats]{dc_oai_listmetadataformats()}} - List metadata formats
\item \code{\link[=dc_oai_listrecords]{dc_oai_listrecords()}} - List records
\item \code{\link[=dc_oai_listsets]{dc_oai_listsets()}} - List sets
}
}

\section{Search functions (Solr based)}{

\itemize{
\item \code{\link[=dc_search]{dc_search()}} - General search, only returns documents
\item \code{\link[=dc_facet]{dc_facet()}} - Faceting only (w/o general search)
\item \code{\link[=dc_mlt]{dc_mlt()}} - More like this (w/o general search)
\item \code{\link[=dc_stats]{dc_stats()}} - Stats search (w/o general search)
}
}

\section{REST API functions}{

\itemize{
\item \code{\link[=dc_data_center]{dc_data_center()}} - Get a single data center
\item \code{\link[=dc_data_centers]{dc_data_centers()}} - Search data centers
\item \code{\link[=dc_member]{dc_member()}} - Get single DataCite member
\item \code{\link[=dc_members]{dc_members()}} - Search DataCite members
\item \code{\link[=dc_work]{dc_work()}} - Get single works
\item \code{\link[=dc_works]{dc_works()}} - Search works
}
}

\section{Vignettes}{


Coming soon...
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
