% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdatacite-package.R
\docType{package}
\name{rdatacite-package}
\alias{rdatacite}
\alias{rdatacite-package}
\title{DataCite R client.}
\description{
DataCite R client.
}
\section{OAI-PMH functions}{


\itemize{
  \item \code{\link{dc_oai_getrecord}} - Get records
  \item \code{\link{dc_oai_identify}} - identify the OAI-PMH service
  \item \code{\link{dc_oai_listidentifiers}} - List identifiers
  \item \code{\link{dc_oai_listmetadataformats}} - List metadata formats
  \item \code{\link{dc_oai_listrecords}} - List records
  \item \code{\link{dc_oai_listsets}} - List sets
}
}

\section{Search functions}{


\itemize{
  \item \code{\link{dc_search}} - General search, only returns documents
  \item \code{\link{dc_facet}} - Faceting only (w/o general search)
  \item \code{\link{dc_mlt}} - More like this (w/o general search)
  \item \code{\link{dc_stats}} - Stats search (w/o general search)
}
}

\section{Vignettes}{


Coming soon...
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

